+++ src/main/java/com/app/middleware/persistence/dto/UserResidentialAddressDTO.java
@@ -18,8 +18,8 @@ public class UserResidentialAddressDTO {
     private boolean isDeleted = false;
     private String addressType;
     private String userPublicId;
-    private Double lat;
-    private Double lng;
+    private float lat;
+    private float lng;
     private String addressLine;
     private String apartmentAddress;
     private String country;
